backend/requirements/base.txt
django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
psycopg2-binary==2.9.9
redis==5.0.7
Pillow==10.4.0
python-dotenv==1.0.1
celery==5.4.0
django-storages==1.14.4
bcrypt==4.1.3
drf-spectacular==0.27.2
